Enable Py36+libuv+Appveyor
All tests are passing locally after fixing a GC-cycle related issue in
the io watcher that kept the wrong socket watchers alive. This should
probably be generalized to all watchers on all FFI platforms.

+++ src/greentest/test___monkey_patching.py
@@ -44,7 +44,7 @@ def TESTRUNNER(tests=None):
                'timeout': TIMEOUT,
                'setenv': {'PYTHONPATH': PYTHONPATH}}
 
-    if tests:
+    if tests and not sys.platform.startswith("win"):
         atexit.register(os.system, 'rm -f */@test*')
 
     basic_args = [sys.executable, '-u', '-W', 'ignore', '-m' 'monkey_test']
